Finish expand fn and move to core.
