+++ sotodlib/preprocess/core.py
@@ -2,6 +2,7 @@
 import logging
 import numpy as np
 from .. import core
+from so3g.proj import RangesMatrix
 
 class _Preprocess(object):
     """The base class for Preprocessing modules which defines the required
@@ -127,7 +128,79 @@ class _Preprocess(object):
                 f"Preprocess Module of name {name} is already Registered"
             )
 
+def expand_proc_aman(calc_aman, proc_aman, flag_fill_val=False,
+                     flt_fill_val=np.nan, int_fill_val=2**32-1):
+    """
+    Helper function to expand the size of calculated products to match the 
+    precut dets and samps shape.
+    Arguments:
+    ----------
+    calc_aman: AxisManager
+        AxisManager returned from calc step.
+    proc_aman: AxisManager
+        AxisManager passed between processes.
+    fill_value:
+        Value to fill missing indices with.
+    """
+    if (proc_aman.dets.count < proc_aman.fdets.count) or \
+       (proc_aman.samps.count < proc_aman.fsamps.count):
+        _, mdets, fmdets = np.intersect1d(proc_aman.dets.vals,
+                                          proc_aman.fdets.vals,
+                                          return_indices=True)
+        fmsamps = slice(proc_aman.samps.offset,
+                        proc_aman.samps.offset+proc_aman.samps.count)
+    
+    out_aman = core.AxisManager(proc_aman.fdets, proc_aman.fsamps)
+    for fld in calc_aman._fields:
+        dat = calc_aman[fld]
+        dax, sax = None, None
+        axes = calc_aman.shape_str(fld).split(',')
+        maxes = np.isin(axes, ['dets', 'samps'])
+        axismap = [(i, a) for i, a in enumerate(axes)]
+        faxismap = [(i, 'f'+a) if m else (i, a) for i, [a, m] in enumerate(list(zip(axes, maxes)))]
+        if np.all(~maxes):
+            out_aman.wrap(fld, dat, axismap)
+        else:
+            if np.isin(type(np.hstack(dat)[0]), [so3g.RangesInt32, bool]):
+                fill_value = flag_fill_val
+                if isinstance(dat, RangesMatrix):
+                    dat = dat.mask()
+            if np.isin(type(np.hstack(dat)[0]), [np.int64, np.int32, int]):
+                fill_value = int_fill_val
+            if np.isin(type(np.hstack(dat)[0]), [np.float64, np.float32, float]):
+                fill_value = flt_fill_val
 
+            shape = tuple([proc_aman[am].count if m else calc_aman[fld].count for am, m in zip(axismap, maxes)])
+            fdat = np.full(shape, fill_value)
+
+            if ('dets' in axes) and not('samps' in axes):
+                dax = np.where(axes == 'dets')[0]
+                dat = np.moveaxis(dat, dax, 0)
+                fdat = np.moveaxis(fdat, dax, 0) 
+                fdat[fmdets] = dat
+                dat = np.moveaxis(dat, 0, dax)
+                fdat = np.moveaxis(fdat, 0, dax)
+            elif ('samps' in axes) and not('dets' in axes):
+                sax = np.where(axes == 'samps')[0]
+                dat = np.moveaxis(dat, sax, 0)
+                fdat = np.moveaxis(fdat, sax, 0)
+                fdat[fmsamps] = dat
+                dat = np.moveaxis(dat, 0, sax)
+                fdat = np.moveaxis(fdat, 0, sax)
+            else:
+                dax = np.where(axes == 'dets')[0]
+                sax = np.where(axes == 'samps')[0]
+                dat = np.moveaxis(dat, (dax, sax), (0,1))
+                fdat = np.moveaxis(fdat, (dax, sax), (0,1))
+                fdat[fmdets, fmsamps] = dat
+                dat = np.moveaxis(dat, (0,1), dax, sax)
+                fdat = np.moveaxis(fdat, (0,1), (dax, sax))
+            
+            if isinstance(fill_value, bool):
+                fdat = RangesMatrix.from_mask(fdat)
+                dat = RangesMatrix.from_mask(dat)
+            out_aman.wrap(fld, dat, axismap)
+            out_aman.wrap(fld, fdat, faxismap)
 
 class Pipeline(list):
     """This class is designed to create and run pipelines out of a series of
